API

PiecewiseLinearFunctions.PiecewiseLinearFunction

PiecewiseLinearFunctions.PiecewiseLinearFunction

PiecewiseLinearFunctions.compose

PiecewiseLinearFunctions.compute_slopes

PiecewiseLinearFunctions.convex_lower_bound

PiecewiseLinearFunctions.convex_meet

PiecewiseLinearFunctions.is_convex

PiecewiseLinearFunctions.is_convex_sequence

PiecewiseLinearFunctions.old_convex_meet

PiecewiseLinearFunctions.remove_redundant_breakpoints

PiecewiseLinearFunctions.PiecewiseLinearFunction — Type

struct PiecewiseLinearFunction{T<:AbstractFloat}

Fields

x::Vector{T} where T<:AbstractFloat: x coordinates of the breakpoints

y::Vector{T} where T<:AbstractFloat: y coordinates of the breakpoints

left_slope::AbstractFloat: slope of the function to the left of the first breakpoint

right_slope::AbstractFloat: slope of the function to the right of the last breakpoint

PiecewiseLinearFunctions.PiecewiseLinearFunction — Method

Evaluate the piecewise linear function f at x.

PiecewiseLinearFunctions.compose — Method

compose(
    f::PiecewiseLinearFunction,
    g::PiecewiseLinearFunction;
    postprocess_breakpoints
) -> PiecewiseLinearFunction

Compute the composition f ∘ g of two piecewise linear functions. By default, the function checks for and removes redundant breakpoints in the output as a post processing step. For faster computation, set postprocess_breakpoints=false.

This method can also be called using Base.∘

PiecewiseLinearFunctions.compute_slopes — Method

compute_slopes(
    f::PiecewiseLinearFunction
) -> Vector{<:AbstractFloat}

Compute all the slopes of the piecewise linear function, from left to right.

PiecewiseLinearFunctions.convex_lower_bound — Method

convex_lower_bound(
    f::PiecewiseLinearFunction{T}
) -> PiecewiseLinearFunction

Compute the tightest convex lower bound of a piecewise linear function, using a convex hull algorithm.

PiecewiseLinearFunctions.convex_meet — Method

convex_meet(
    f::PiecewiseLinearFunction,
    g::PiecewiseLinearFunction
) -> PiecewiseLinearFunction

A faster version of convex_meet that only works on convex functions.

Warning

This function does not check if the input functions are convex.

PiecewiseLinearFunctions.is_convex — Method

is_convex(f::PiecewiseLinearFunction; eps) -> Any

Check if a piecewise linear function is convex.

PiecewiseLinearFunctions.is_convex_sequence — Method

is_convex_sequence(f, i1, i2, i3) -> Any

Check if the sequence of points (p1, p2, p3) forms a convex sequence.

PiecewiseLinearFunctions.old_convex_meet — Method

old_convex_meet(
    f::PiecewiseLinearFunction{T},
    g::PiecewiseLinearFunction{T}
) -> PiecewiseLinearFunction

Compute the convex meet of two piecewise linear functions, i.e. the tightest convex lower bound.

PiecewiseLinearFunctions.remove_redundant_breakpoints — Method

remove_redundant_breakpoints(
    f::PiecewiseLinearFunction;
    atol
) -> PiecewiseLinearFunction

Return a new piecewise linear function with redundant breakpoints removed.