API

PiecewiseLinearFunctions.PiecewiseLinearFunction

PiecewiseLinearFunctions.PiecewiseLinearFunction

PiecewiseLinearFunctions.compose

PiecewiseLinearFunctions.compute_slopes

PiecewiseLinearFunctions.convex_meet

PiecewiseLinearFunctions.is_convex

PiecewiseLinearFunctions.remove_redundant_breakpoints

PiecewiseLinearFunctions.PiecewiseLinearFunction — Type

struct PiecewiseLinearFunction{T<:AbstractFloat}

Fields

x::Vector{T} where T<:AbstractFloat: x coordinates of the breakpoints

y::Vector{T} where T<:AbstractFloat: y coordinates of the breakpoints

left_slope::AbstractFloat: slope of the function to the left of the first breakpoint

right_slope::AbstractFloat: slope of the function to the right of the last breakpoint

PiecewiseLinearFunctions.PiecewiseLinearFunction — Method

Evaluate the piecewise linear function f at x.

PiecewiseLinearFunctions.compose — Method

compose(
    f::PiecewiseLinearFunction{T},
    g::PiecewiseLinearFunction{T};
    postprocess_breakpoints
) -> PiecewiseLinearFunction

Compute the composition f ∘ g of two piecewise linear functions. By default, the function removes redundant breakpoints in the output as a post processing step.

This method can also be called using Base.∘

PiecewiseLinearFunctions.compute_slopes — Method

compute_slopes(
    f::PiecewiseLinearFunction
) -> Vector{<:AbstractFloat}

Compute all the slopes of the piecewise linear function, from left to right.

PiecewiseLinearFunctions.convex_meet — Method

convex_meet(
    f::PiecewiseLinearFunction{T},
    g::PiecewiseLinearFunction{T}
) -> PiecewiseLinearFunction

Compute the convex meet of two piecewise linear functions, i.e. the tightest convex lower bound. Both functions need to be convex.

PiecewiseLinearFunctions.is_convex — Method

is_convex(f::PiecewiseLinearFunction) -> Any

Check if a piecewise linear function is convex.

PiecewiseLinearFunctions.remove_redundant_breakpoints — Method

remove_redundant_breakpoints(
    f::PiecewiseLinearFunction;
    atol
) -> PiecewiseLinearFunction

Return a new piecewise linear function with redundant breakpoints removed.